try limiting test amounts again, but three files this time [ci skip]
Last time I pushed just the test_editor file and the tests started
passing. And I can't reproduce locally. So if I'm going to be debugging
on CI it would be nice if the test would take less time to run.

all features: editor scenarios fails
just editor.feature: passes
editor.feature and downloads.feature (the one before it in the log): passes
editor.feature and downloads.feature and hints.feature (the one after it): ???