Try clearing messages before running tests? [ci skip]
I tried to make the test runs faster by only running a subset of the
tests, but they always pass when I do that ;_;

I've got a screenshot and it shows many many messages on the screen,
mostly from the download tests it seems. It also seems to show the
jsprompt file behind since I can see the two "Show prompt" and "Show
prompt with default value" buttons. And that is the last page that the
download tests beforehand open, mayb it's stuck on that page? I do see a
"No folder chosen" on the screen, but I can't see the whole screen yet.
Gotta re-run with the geometry fix.

If this doesn't work I guess I'll make sure it has a fresh window?