Exit mode before edit tests.
The screenshots help! Wow!

The screenshots appear to show that the prompt from the jsprompt page
which is the last one that the downloads.feature tests open.

Does having a prompt open prevent you from opening a new URL?
Why does this not reproduce when I run {downloads,editor,hints} but only
with the whole test suite?

These things I don't know. I just push and wait!

I'm exiting prompt mode here. It would probably be better to clear the
mode in the next test file, but when I try to do that and it's not in
prompt mode for that one it fails the test because of a command error.
I did try "Given I have a fresh instance" but that makes the
editor.feature test file take over twice as long! Does it get a fresh
instance before every test? Aren't we supposed to get a new one at the
start of a scenario?