try macos 14 M1 in GHA
There is reports here that the renderer processes are crashing on macos
14 on apple silicon: https://github.com/qutebrowser/qutebrowser/discussions/8226

It seems to work fine when the .dmg is installed directly on macos 11.
Homebrew seems to be just installing the .dmg and doing nothing else? https://github.com/Homebrew/homebrew-cask/blob/master/Casks/q/qutebrowser.rb

So lets see what happens if we try to build on the M1 github actions
runners.