Run tests in parallel on a schedule.
So we can see how reliable they are compared to the non parallel runs.
And how much faster they are.

TODO: keep this branch up to data so it gets fixes for flaky tests etc.
Maybe I should have to scheduled actions which check out the upstream
repo and run the tests with different args.